Keep consistent system test names
